handle installing Nomad into a Laravel app
As `illuminate/foundation` doesn't exist as a standalone package, we
leverage `laravel-zero/foundation` in order to be able to work with
migrations outside of the context of a Laravel app.

When we install into a Laravel application, however, this dependency
leads to a) namespace collisions and b) Composer aiming to use the
incorrect dependency.

We introduce a `pre-autoload-dump` script, which should handle detecting
if we're in a Laravel application and removing the offending package